+++ src/qt/chaincoin.cpp
@@ -37,9 +37,6 @@
 #include <util.h>
 #include <warnings.h>
 
-#ifdef ENABLE_WALLET
-#include <wallet/init.h>
-#endif
 #include <walletinitinterface.h>
 
 #include <memory>
@@ -670,11 +667,6 @@ int main(int argc, char *argv[])
     // Start up the payment server early, too, so impatient users that click on
     // chaincoin: links repeatedly have their payment requests routed to this process:
     app.createPaymentServer();
-
-    // Hook up the wallet interface
-    g_wallet_init_interface.reset(new WalletInit);
-#else
-    g_wallet_init_interface.reset(new DummyWalletInit);
 #endif
 
     /// 9. Main GUI initialization
