+++ src/wallet/init.cpp
@@ -3,8 +3,6 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <wallet/init.h>
-
 #include <chainparams.h>
 #include <init.h>
 #include <interface/moduleinterface.h>
@@ -14,9 +12,45 @@
 #include <validation.h>
 #include <wallet/keepass.h>
 #include <wallet/privatesend-client.h>
+#include <walletinitinterface.h>
 #include <wallet/rpcwallet.h>
 #include <wallet/wallet.h>
 #include <wallet/walletutil.h>
+
+class WalletInit : public WalletInitInterface {
+public:
+
+    //! Return the wallets help message.
+    std::string GetHelpString(bool showDebug) override;
+
+    //! Wallets parameter interaction
+    bool ParameterInteraction() override;
+
+    //! Register wallet RPCs.
+    void RegisterRPC(CRPCTable &tableRPC) override;
+
+    //! Responsible for reading and validating the -wallet arguments and verifying the wallet database.
+    //  This function will perform salvage on the wallet if requested, as long as only one wallet is
+    //  being loaded (WalletParameterInteraction forbids -salvagewallet, -zapwallettxes or -upgradewallet with multiwallet).
+    bool Verify() override;
+
+    //! Load wallet databases.
+    bool Open() override;
+
+    //! Complete startup of wallets.
+    void Start(CScheduler& scheduler, CConnman* connman) override;
+
+    //! Flush all wallets in preparation for shutdown.
+    void Flush() override;
+
+    //! Stop all wallets. Wallets will be flushed first.
+    void Stop() override;
+
+    //! Close all wallets.
+    void Close() override;
+};
+
+std::unique_ptr<WalletInitInterface> g_wallet_init_interface(new WalletInit);
 
 std::string WalletInit::GetHelpString(bool showDebug)
 {
