+++ src/wallet/wallet.cpp
@@ -14,7 +14,6 @@
 #include <consensus/consensus.h>
 #include <consensus/validation.h>
 #include <fs.h>
-#include <wallet/init.h>
 #include <key.h>
 #include <key_io.h>
 #include <keystore.h>
