+++ src/walletinitinterface.h
@@ -35,18 +35,4 @@ public:
     virtual ~WalletInitInterface() {}
 };
 
-class DummyWalletInit : public WalletInitInterface {
-public:
-
-    std::string GetHelpString(bool showDebug) override {return std::string{};}
-    bool ParameterInteraction() override {return true;}
-    void RegisterRPC(CRPCTable &) override {}
-    bool Verify() override {return true;}
-    bool Open() override {return true;}
-    void Start(CScheduler& scheduler, CConnman* connman) override {}
-    void Flush() override {}
-    void Stop() override {}
-    void Close() override {}
-};
-
 #endif // BITCOIN_WALLETINITINTERFACE_H
